refactor: rename AboutPage to About
